Fix macosx default system font (#17)
* Fix macosx default system font

- ".AppleSystemUIFont" does not return a valid CGFontRef when used as a font name!
        - We need to go through NSFont and get the apt CGFontRef
- Along with this, when creating hb_font using hb's coretext api on the above CGFontRef, it seems to
be missing MORT and MORX font tables, which are necessary and required in hb's shape planner code.
Because of this limitation, I ended up going the route of directly creating hb_font using ft_Face,
which inturn is created from raw data read from the CGFontRef. Since we are not concerned about
shaping with this "LATIN" font, we do not have to worry about creating a harfbuzz context with
coretext for this font.

* Uses newer coretext hbfont create API

- updates harfbuzz dependency requirement of 1.7.2 for OSX
- Using harfbuzz 1.7.2 and above resolves an issue with defautl osx system font, as this version of
harfbuzz works off kern table present in this font.